+++ CHANGELOG.md
@@ -35,6 +35,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
   - `client.batch.create(entities, options)`
   - `client.batch.update(entities, options)`
   - `client.batch.deleteById(ids, options)`
+  - `client.batch.deleteByFilter(filter, limit, options)`
 
 ### CHANGED
 - Client no longer throws a Axios error if a response has been received and response is a Haystack of GraphQL error.